sq set project key dynamically and add new stage for fork

+++ src/org/ods/quickstarter/PushToRemoteStage.groovy
@@ -16,29 +16,38 @@ class PushToRemoteStage extends Stage {
     }
 
     script.dir(context.targetDir) {
+      if (!script.fileExists (".git")) {
+        script.echo("Initializing quickstarter git repo ${context.targetDir} @${context.gitUrlHttp}")
+        script.sh(
+          script: """
+          # Clone first (there is ALWAYS a remote repo!)
+          git clone ${context.gitUrlHttp}
+          basename=\$(basename ${context.gitUrlHttp})
+          clonedGitFolderName=\${basename%.*}
+          files=\$(ls \$clonedGitFolderName | grep -v ^l | wc -l)
+          if [ \$files == 0 ]; then 
+            echo "Init project from scratch \$(pwd)"
+            git init .
+            git remote add origin ${context.gitUrlHttp}
+          else
+            echo "Upgrading existing project: ${context.gitUrlHttp}"
+            mv "\$clonedGitFolderName"/* .
+            mv "\$clonedGitFolderName"/.[!.]* .
+          fi
+          rm -rf \$clonedGitFolderName
+          git config user.email "undefined"
+          git config user.name "ODS System User"
+          git add --all .
+          git commit -m "Initial OpenDevStack commit"
+          """,
+          label: "Copy and commit quickstarter files"
+        )
+      }
+      script.echo("Pushing quickstarter git repo to ${context.gitUrlHttp}")
       script.sh(
         script: """
-        # Clone first (there is ALWAYS a remote repo!)
-        git clone ${context.gitUrlHttp}
-        basename=\$(basename ${context.gitUrlHttp})
-        clonedGitFolderName=\${basename%.*}
-        files=\$(ls \$clonedGitFolderName | grep -v ^l | wc -l)
-        if [ \$files == 0 ]; then 
-          echo "Init project from scratch \$(pwd)"
-          git init .
-          git remote add origin ${context.gitUrlHttp}
-        else
-          echo "Upgrading existing project: ${context.gitUrlHttp}"
-          mv "\$clonedGitFolderName"/* .
-          mv "\$clonedGitFolderName"/.[!.]* .
-        fi
-        git config user.email "undefined"
-        git config user.name "ODS System User"
-        rm -rf \$clonedGitFolderName
-        git add --all .
-        git commit -m "Initial OpenDevStack commit"
-        git push -u origin ${config.branch}
-        """,
+          git push -u origin ${config.branch}
+          """,
         label: "Push to remote"
       )
     }
